moved api.js to utils folder
